# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookstore API",
        "description": "API documentation for Bookstore",
        "version": "1.0.0"
    },
    "paths": {
        "/auth/send-code": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Send code to user",
                "description": "Send code to user",
                "operationId": "47e22916ffd4cbf8a6081f5421405b99",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "09123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "send code"
                    },
                    "401": {
                        "description": "error"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "login user",
                "description": "Login user and return token",
                "operationId": "a2a122b9d71e1e1afbfb813ed4790930",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "code"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "09123456789"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "login"
                    },
                    "401": {
                        "description": "Error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        }
    ]
}